app: Add new scan/result endpoint
In the process of using the new, SCAN-specific error template page, it
became apparent that the conditional source ( source == 'scan' ?) logic
for rendering a different results page for SCAN would become
increasingly hairy for rendering a different error page, too.

The home page for SCAN already lives at a separate endpoing (`/scan`).
To simplify things, move the SCAN-specific `/result` logic to a new
endpoint, `/scan/result`.

This commit contains a bit of a hack workaround for local development
that I would like to improve eventually, but I'm including it now for
ease's sake.