Update README.md
Simplify the instructions for how to start the app by splitting the
installation and actually running of the app. Simplify further by
providing the `FLASK_ENV` environment variable at runtime rather than
exporting it.

One of the members of our lab was having difficulty running the app
while copy/pasting the instructions as they were previously written, so
hopefully this makes things easier.